+++ src/components/shared/index.ts
@@ -1,3 +1,4 @@
+export * from './BusySplash';
 export * from './ContentFooter';
 export * from './EmptyPlaceholder';
 export * from './ErrorPlaceHolder';